WCM-573 Move methods to base class for reuse
